add float field for coins model

+++ radarcxapp/models.py
@@ -5,7 +5,7 @@ from django.utils import timezone
 
 class Coin(models.Model):
     name = models.CharField(max_length=20)
-    realtime_price = models.IntegerField(default=0)
+    realtime_price = models.FloatField(default=0)
     moving_average = models.IntegerField(default=0)
     volume = models.IntegerField(default=0)
     date_updated = models.DateTimeField(default=timezone.now)
